Rename ecs package to entities
