+++ internal/systems/movement.go
@@ -4,12 +4,12 @@ import (
 	"reflect"
 
 	"github.com/juanancid/maze-adventure/internal/components"
-	"github.com/juanancid/maze-adventure/internal/ecs"
+	"github.com/juanancid/maze-adventure/internal/entities"
 )
 
 type Movement struct{}
 
-func (ms *Movement) Update(w *ecs.World) {
+func (ms *Movement) Update(w *entities.World) {
 	for entity, entityVelocity := range w.GetComponents(reflect.TypeOf(&components.Velocity{})) {
 		velocity := entityVelocity.(*components.Velocity)
 		position := w.GetComponent(entity, reflect.TypeOf(&components.Position{})).(*components.Position)
